[workerd-cxx] require lifetime annotations on all returned references (#3)
To support error propagation cxx generates C interface
where successful result value is passed out as a pointer.
This requires moving result type to parameter position
which changes the meaning according to the rules of
implicit liftime (each parameter gets its own lifetime).

Rather than implement the same set of rules that compiler
to generate such annotations, demand the user to always
have explicit lifetimes for returned references.

This should not be a great deal in practice with
`unsafe` being the only inconvenience.

In return the follow up no-abort work gets much simpler.
I might revisit this change in the future if the burden
of explicit lifetimes becomes too high.

[workerd-cxx] fixing build (#6)