[workerd-cxx] Add extensive testing and refactor testing
Additionally, renamed KjOwn to Own to be more in line with standard Rust smart pointer conventions.
Test refactor was done by hand, but the additional tests were scaffolded by claude code.
Some generated tests would obviously not compile due to the borrow checker, which seemed to be a major
overlook by claude. Other than that, the other generated tests seemed to be reasonable with a few tweaks,
and signifigantly boost my confidence that moving and borrowing function correctly with Own.